backend/requirements.txt
Flask==2.1.0
Werkzeug==2.1.0
psycopg2-binary==2.9.3
requests==2.31.0
